Update multi-RFP request and service
This commit updates the `UpdateMultiRFPRequest` class in `update-multi-RFP.request.ts` to replace the `time_type_id` property with `preferred_testing_time` as an array. It also updates the `MultiRfpService` class in `multi-rfp.service.ts` to uncomment the `@Inject` decorator for `RequestForProposalService` and to uncomment the line that updates the request for proposal.